extract MultipartRequest and MapField
