+++ apollo-router/src/plugins/file_uploads/mod.rs
@@ -1,34 +1,16 @@
-use std::collections::HashMap;
-use std::collections::HashSet;
-use std::mem;
 use std::ops::ControlFlow;
-use std::pin::Pin;
 use std::sync::Arc;
-use std::task;
-use std::task::Poll;
 
-use bytes::Bytes;
 use futures::FutureExt;
-use futures::Stream;
 use http::header::CONTENT_LENGTH;
 use http::header::CONTENT_TYPE;
-use http::HeaderMap;
 use http::HeaderName;
 use http::HeaderValue;
-use indexmap::IndexMap;
-use indexmap::IndexSet;
-use itertools::Itertools;
 use mediatype::names::BOUNDARY;
 use mediatype::names::FORM_DATA;
 use mediatype::names::MULTIPART;
 use mediatype::MediaType;
 use mediatype::ReadParams;
-use multer::Constraints;
-use multer::Multipart;
-use multer::SizeLimit;
-use pin_project_lite::pin_project;
-use tokio::sync::Mutex;
-use tokio::sync::OwnedMutexGuard;
 use tower::BoxError;
 use tower::ServiceBuilder;
 use tower::ServiceExt;
@@ -36,7 +18,10 @@ use tower::ServiceExt;
 use self::config::FileUploadsConfig;
 use self::config::MultipartRequestLimits;
 use self::error::FileUploadError;
+use self::map_field::MapField;
+use self::map_field::MapFieldRaw;
 use self::multipart_form_data::MultipartFormData;
+use self::multipart_request::MultipartRequest;
 use self::rearrange_query_plan::rearange_query_plan;
 use crate::layers::ServiceBuilderExt;
 use crate::plugin::PluginInit;
@@ -49,14 +34,12 @@ use crate::services::supergraph;
 
 mod config;
 mod error;
+mod map_field;
 mod multipart_form_data;
+mod multipart_request;
 mod rearrange_query_plan;
 
 type UploadResult<T> = Result<T, error::FileUploadError>;
-
-// The limit to set for the map field in the multipart request.
-// We don't expect this to ever be reached, but we can always add a config option if needed later.
-const MAP_SIZE_LIMIT: u64 = 10 * 1024;
 
 // FIXME: check if we need to hide docs
 #[doc(hidden)] // Only public for integration tests
@@ -267,14 +250,10 @@ fn try_path<'a>(
     })
 }
 
-type MapField = IndexMap<String, Vec<String>>;
-type MapPerVariable = HashMap<String, MapPerFile>;
-type MapPerFile = HashMap<String, Vec<Vec<String>>>;
-
 #[derive(Clone)]
 struct SupergraphLayerResult {
     multipart: MultipartRequest,
-    map: Arc<ParsedMap>,
+    map: Arc<MapField>,
 }
 
 fn execution_layer(req: execution::Request) -> UploadResult<execution::Request> {
@@ -301,30 +280,12 @@ async fn subgraph_layer(mut req: subgraph::Request) -> subgraph::Request {
         .get::<SupergraphLayerResult>()
         .cloned();
     if let Some(supergraph_result) = supergraph_result {
-        let SupergraphLayerResult {
-            multipart,
-            map,
-        } = supergraph_result;
+        let SupergraphLayerResult { multipart, map } = supergraph_result;
 
         let variables = &mut req.subgraph_request.body_mut().variables;
-        let mut map_field: MapField = IndexMap::new();
         for (variable_name, variable_value) in variables.iter_mut() {
-            let variable_name = variable_name.as_str();
-            if let Some(variable_map) = map.map_per_variable.get(variable_name) {
-                for (file, paths) in variable_map.iter() {
-                    map_field.insert(
-                        file.clone(),
-                        paths
-                            .iter()
-                            .map(|path| {
-                                if path.is_empty() {
-                                    format!("variables.{}", variable_name)
-                                } else {
-                                    format!("variables.{}.{}", variable_name, path.join("."))
-                                }
-                            })
-                            .collect(),
-                    );
+            if let Some(variable_map) = map.map_per_variable.get(variable_name.as_str()) {
+                for paths in variable_map.values() {
                     for path in paths {
                         if let Some(json_value) = try_path(variable_value, path) {
                             json_value.take();
@@ -333,13 +294,14 @@ async fn subgraph_layer(mut req: subgraph::Request) -> subgraph::Request {
                 }
             }
         }
-        if !map_field.is_empty() {
-            map_field.sort_by_cached_key(|file, _| map.files_order.get_index_of(file));
+
+        let subgraph_map = map.sugraph_map(variables.keys());
+        if !subgraph_map.is_empty() {
             req.subgraph_request
                 .extensions_mut()
                 .insert(SubgraphHttpRequestExtensions {
                     multipart,
-                    map_field,
+                    map_field: subgraph_map,
                 });
         }
     }
@@ -348,7 +310,7 @@ async fn subgraph_layer(mut req: subgraph::Request) -> subgraph::Request {
 
 struct SubgraphHttpRequestExtensions {
     multipart: MultipartRequest,
-    map_field: MapField,
+    map_field: MapFieldRaw,
 }
 
 static APOLLO_REQUIRE_PREFLIGHT: HeaderName = HeaderName::from_static("apollo-require-preflight");
@@ -378,291 +340,4 @@ pub(crate) async fn http_request_wrapper(
         );
     }
     req
-}
-
-#[derive(Clone)]
-struct MultipartRequest {
-    state: Arc<Mutex<MultipartRequestState>>,
-}
-
-struct MultipartRequestState {
-    multer: multer::Multipart<'static>,
-    limits: MultipartRequestLimits,
-    read_files_counter: usize,
-    file_sizes: Vec<usize>,
-    max_files_exceeded: bool,
-    max_files_size_exceeded: bool,
-}
-
-impl Drop for MultipartRequestState {
-    fn drop(&mut self) {
-        u64_counter!(
-            "apollo.router.operations.file_uploads",
-            "file uploads",
-            1,
-            "file_uploads.limits.max_file_size.exceeded" = self.max_files_size_exceeded,
-            "file_uploads.limits.max_files.exceeded" = self.max_files_exceeded
-        );
-
-        for file_size in &self.file_sizes {
-            u64_histogram!(
-                "apollo.router.operations.file_uploads.file_size",
-                "file upload sizes",
-                (*file_size) as u64
-            );
-        }
-        u64_histogram!(
-            "apollo.router.operations.file_uploads.files",
-            "number of files per request",
-            self.read_files_counter as u64
-        );
-    }
-}
-
-impl MultipartRequest {
-    fn new(request_body: hyper::Body, boundary: String, limits: MultipartRequestLimits) -> Self {
-        let multer = Multipart::with_constraints(
-            request_body,
-            boundary,
-            Constraints::new().size_limit(SizeLimit::new().for_field("map", MAP_SIZE_LIMIT)),
-        );
-        Self {
-            state: Arc::new(Mutex::new(MultipartRequestState {
-                multer,
-                limits,
-                read_files_counter: 0,
-                file_sizes: Vec::new(),
-                max_files_exceeded: false,
-                max_files_size_exceeded: false,
-            })),
-        }
-    }
-
-    async fn operations_field(&mut self) -> UploadResult<multer::Field<'static>> {
-        self.state
-            .lock()
-            .await
-            .multer
-            .next_field()
-            .await?
-            .filter(|field| field.name() == Some("operations"))
-            .ok_or_else(|| FileUploadError::MissingOperationsField)
-    }
-
-    async fn map_field(&mut self) -> UploadResult<ParsedMap> {
-        let mut state = self.state.lock().await;
-        let bytes = state
-            .multer
-            .next_field()
-            .await?
-            .filter(|field| field.name() == Some("map"))
-            .ok_or_else(|| FileUploadError::MissingMapField)?
-            .bytes()
-            .await?;
-
-        let map_field: MapField =
-            serde_json::from_slice(&bytes).map_err(FileUploadError::InvalidJsonInMapField)?;
-
-        let limit = state.limits.max_files;
-        if map_field.len() > limit {
-            state.max_files_exceeded = true;
-            return Err(FileUploadError::MaxFilesLimitExceeded(limit));
-        }
-        ParsedMap::new(map_field)
-    }
-
-    async fn subgraph_stream<FilePrefixFn>(
-        &mut self,
-        file_names: HashSet<String>,
-        file_prefix_fn: FilePrefixFn,
-    ) -> SubgraphFileProxyStream<FilePrefixFn>
-    where
-        FilePrefixFn: Fn(&HeaderMap) -> Bytes,
-    {
-        let state = self.state.clone().lock_owned().await;
-        SubgraphFileProxyStream::new(state, file_names, file_prefix_fn)
-    }
-}
-
-struct ParsedMap {
-    files_order: IndexSet<String>,
-    map_per_variable: MapPerVariable,
-}
-
-impl ParsedMap {
-    fn new(map_field: MapField) -> UploadResult<Self> {
-        let mut files_order = IndexSet::new();
-        let mut map_per_variable: MapPerVariable = HashMap::new();
-        for (filename, paths) in map_field.into_iter() {
-            for path in paths.into_iter() {
-                let mut segments = path.split('.');
-                let first_segment = segments.next();
-                if first_segment != Some("variables") {
-                    if first_segment
-                        .and_then(|str| str.parse::<usize>().ok())
-                        .is_some()
-                    {
-                        return Err(FileUploadError::BatchRequestAreNotSupported);
-                    }
-                    return Err(FileUploadError::InvalidPathInsideMapField(path));
-                }
-                let variable_name = segments.next().ok_or_else(|| {
-                    FileUploadError::MissingVariableNameInsideMapField(path.clone())
-                })?;
-                let variable_path: Vec<String> = segments.map(str::to_owned).collect();
-
-                map_per_variable
-                    .entry(variable_name.to_owned())
-                    .or_default()
-                    .entry(filename.clone())
-                    .or_default()
-                    .push(variable_path);
-            }
-            files_order.insert(filename);
-        }
-
-        Ok(Self {
-            files_order,
-            map_per_variable,
-        })
-    }
-}
-
-pin_project! {
-    struct SubgraphFileProxyStream<FilePrefixFn> {
-        state: OwnedMutexGuard<MultipartRequestState>,
-        file_names: HashSet<String>,
-        file_prefix_fn: FilePrefixFn,
-        #[pin]
-        current_field: Option<multer::Field<'static>>,
-        current_field_bytes: usize,
-    }
-}
-
-impl<FilePrefixFn> SubgraphFileProxyStream<FilePrefixFn>
-where
-    FilePrefixFn: Fn(&HeaderMap) -> Bytes,
-{
-    fn new(
-        state: OwnedMutexGuard<MultipartRequestState>,
-        file_names: HashSet<String>,
-        file_prefix_fn: FilePrefixFn,
-    ) -> Self {
-        Self {
-            state,
-            file_names,
-            file_prefix_fn,
-            current_field: None,
-            current_field_bytes: 0,
-        }
-    }
-
-    fn poll_current_field(
-        &mut self,
-        cx: &mut task::Context<'_>,
-    ) -> Poll<Option<UploadResult<Bytes>>> {
-        if let Some(field) = &mut self.current_field {
-            let filename = field
-                .file_name()
-                .or_else(|| field.name())
-                .map(|name| format!("'{}'", name))
-                .unwrap_or_else(|| "unknown".to_owned());
-
-            let field = Pin::new(field);
-            match field.poll_next(cx) {
-                Poll::Pending => Poll::Pending,
-                Poll::Ready(None) => {
-                    self.current_field = None;
-                    let file_size = self.current_field_bytes;
-                    self.state.file_sizes.push(file_size);
-                    Poll::Ready(None)
-                }
-                Poll::Ready(Some(Ok(bytes))) => {
-                    self.current_field_bytes += bytes.len();
-                    let limit = self.state.limits.max_file_size;
-                    if self.current_field_bytes > (limit.as_u64() as usize) {
-                        self.current_field = None;
-                        self.state.max_files_size_exceeded = true;
-                        Poll::Ready(Some(Err(FileUploadError::MaxFileSizeLimitExceeded {
-                            limit,
-                            filename,
-                        })))
-                    } else {
-                        Poll::Ready(Some(Ok(bytes)))
-                    }
-                }
-                Poll::Ready(Some(Err(e))) => {
-                    Poll::Ready(Some(Err(FileUploadError::InvalidMultipartRequest(e))))
-                }
-            }
-        } else {
-            Poll::Ready(None)
-        }
-    }
-
-    fn poll_next_field(
-        mut self: Pin<&mut Self>,
-        cx: &mut task::Context<'_>,
-    ) -> Poll<Option<UploadResult<Bytes>>> {
-        if self.file_names.is_empty() {
-            return Poll::Ready(None);
-        }
-        loop {
-            match self.state.multer.poll_next_field(cx) {
-                Poll::Pending => return Poll::Pending,
-                Poll::Ready(Ok(None)) => {
-                    if self.file_names.is_empty() {
-                        return Poll::Ready(None);
-                    }
-
-                    let files = mem::take(&mut self.file_names);
-                    return Poll::Ready(Some(Err(FileUploadError::MissingFiles(
-                        files
-                            .into_iter()
-                            .map(|file| format!("'{}'", file))
-                            .join(", "),
-                    ))));
-                }
-                Poll::Ready(Ok(Some(field))) => {
-                    let limit = self.state.limits.max_files;
-                    if self.state.read_files_counter == limit {
-                        self.state.max_files_exceeded = true;
-                        return Poll::Ready(Some(Err(FileUploadError::MaxFilesLimitExceeded(
-                            limit,
-                        ))));
-                    } else {
-                        self.state.read_files_counter += 1;
-
-                        if let Some(name) = field.name() {
-                            if self.file_names.remove(name) {
-                                let prefix = (self.file_prefix_fn)(field.headers());
-                                self.current_field = Some(field);
-                                return Poll::Ready(Some(Ok(prefix)));
-                            }
-                        }
-
-                        // The file is extraneous, but the rest can still be processed.
-                        // Just ignore it and don’t exit with an error.
-                        // Matching https://github.com/jaydenseric/graphql-upload/blob/f24d71bfe5be343e65d084d23073c3686a7f4d18/processRequest.mjs#L231-L236
-                    }
-                }
-                Poll::Ready(Err(e)) => return Poll::Ready(Some(Err(e.into()))),
-            }
-        }
-    }
-}
-
-impl<FilePrefixFn> Stream for SubgraphFileProxyStream<FilePrefixFn>
-where
-    FilePrefixFn: Fn(&HeaderMap) -> Bytes,
-{
-    type Item = UploadResult<Bytes>;
-
-    fn poll_next(mut self: Pin<&mut Self>, cx: &mut task::Context<'_>) -> Poll<Option<Self::Item>> {
-        let field_result = self.poll_current_field(cx);
-        match field_result {
-            Poll::Ready(None) => self.poll_next_field(cx),
-            _ => field_result,
-        }
-    }
 }
